Fix data value for cooldown

+++ src/models/Cooldown.js
@@ -3,7 +3,7 @@ const { DataTypes } = require('sequelize');
 const cooldownSchema = {
     id:            { type: DataTypes.STRING, notNull: true, primaryKey: true },
     type:          { type: DataTypes.STRING, notNull: true },
-    expire:        { type: DataTypes.INTEGER, notNull: true }
+    expire:        { type: DataTypes.BIGINT, notNull: true }
 };
